New File structure, added basic instruction time measurement, added 'hardware_emulator' (delay signals), implemented JNI message().

+++ hardware_emulator/src/os_windows/dll_loader.h
@@ -13,7 +13,7 @@ namespace OS {
         ulong base_of_code;
         ulong entry_point;
         ulong section_align;
-        ulong size_of_headers;
+        uint size_of_headers;
         
         void load_values( void *pe );
     };
